scripts/src/requirements.txt
albumentations==1.3.0
numpy==1.23.4
opencv_python_headless==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
scikit_learn==1.1.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
